[rackspace|auto_scale] fixing formatting and documentation; updated collections to pass along dependent parent models

+++ tests/rackspace/models/auto_scale/policy_tests.rb
@@ -4,28 +4,28 @@ Shindo.tests('Fog::Rackspace::AutoScale | policy', ['rackspace', 'rackspace_auto
 
   pending if Fog.mocking?
 
-  group = service.groups.create({
-    :policies => POLICIES_OPTIONS,
-    :group_config => GROUP_CONFIG_OPTIONS,
-    :launch_config => LAUNCH_CONFIG_OPTIONS
-  })
-
-  options = {
-    :name => "policy 2",
-    :change => 5,
-    :cooldown => 100,
-    :type => 'webhook',
-    :group_id => group.id
-  }
-
   begin
+    group = service.groups.create({
+      :policies => POLICIES_OPTIONS,
+      :group_config => GROUP_CONFIG_OPTIONS,
+      :launch_config => LAUNCH_CONFIG_OPTIONS
+    })
+
+    options = {
+      :name => "policy 2",
+      :change => 5,
+      :cooldown => 100,
+      :type => 'webhook',
+      :group => group
+    }
+
     model_tests(group.policies, options, false) do
       tests('#webhooks').succeeds do
         @instance.webhooks
       end
     end
   ensure
-    group.destroy
+    group.destroy if group
   end
 
 end
