+++ tests/rackspace/models/auto_scale/webhook_tests.rb
@@ -4,34 +4,34 @@ Shindo.tests('Fog::Rackspace::AutoScale | webhook', ['rackspace', 'rackspace_aut
 
   pending if Fog.mocking?
 
-  group = service.groups.create({
-    :policies => POLICIES_OPTIONS,
-    :group_config => GROUP_CONFIG_OPTIONS,
-    :launch_config => LAUNCH_CONFIG_OPTIONS
-  })
-
-  policy = group.policies.create({
-    :name => "set group to 5 servers",
-    :desired_capacity => 5,
-    :cooldown => 1800,
-    :type => "webhook",
-    :group_id => group.id
-  })
-
-  options = {
-    :name => 'webhook name',
-    :metadata => {
-      'owner' => 'me'
-    },
-    :group_id => group.id,
-    :policy_id => policy.id
-  }
-
   begin
+    group = service.groups.create({
+      :policies => POLICIES_OPTIONS,
+      :group_config => GROUP_CONFIG_OPTIONS,
+      :launch_config => LAUNCH_CONFIG_OPTIONS
+    })
+
+    policy = group.policies.create({
+      :name => "set group to 5 servers",
+      :desired_capacity => 5,
+      :cooldown => 1800,
+      :type => "webhook",
+      :group => group
+    })
+
+    options = {
+      :name => 'webhook name',
+      :metadata => {
+        'owner' => 'me'
+      },
+      :group => group,
+      :policy => policy
+    }
+
     model_tests(policy.webhooks, options, false)
   ensure
-    policy.destroy
-    group.destroy
+    policy.destroy if policy
+    group.destroy if group
   end
 
 end
